Add yorkie to handle pre-commit gitHook and run eslint
corteza-js don't trigger lint on a commit compare to the other repo.
I  thought it would be usefull for everyone.
I added yorkie, the pkg is already use in the other repo (webapp-* and
corteza-vue, it come as a dependencie of @vue/cli-plugin-eslint) to handle gitHooks.
I also copy the configuration from the other repo to launch lint on
pre-commit